+++ pynapple/core/time_series.py
@@ -453,31 +453,71 @@ class BaseTsd(Base, NDArrayOperatorsMixin, abc.ABC):
 
             return self.__class__(t=time_array, d=new_data_array, time_support=ep)
 
-    def smooth(self, std, size):
-        """Smooth a time series with a gaussian kernel. std is the standard deviation and size is the number of point of the window.
+    def smooth(self, std, time_units="s", size_factor=100, norm=True):
+        """Smooth a time series with a gaussian kernel.
 
-        See the scipy documentation : https://docs.scipy.org/doc/scipy/reference/generated/scipy.signal.windows.gaussian.html
+        `std` is the standard deviation of the gaussian kernel in units of time.
+        If only `std` is passed, the function will compute the standard deviation and size in number
+        of time points automatically based on the sampling rate of the time series.
+        For example, if the time series `tsd` has a sample rate of 100 Hz and `std` is 20 ms,
+        the standard deviation will be converted to an integer through
+        `tsd.rate * std = int(100 * 0.05) = 5`.
+        By default, the function will select a kernel size as 100 times the std in number of time points.
+        `size_factor` can be used to resize the gaussian kernel.
+
+        `norm` set to True normalizes the gaussian kernel to sum to 1.
+
+        In the following example, a time series `tsd` with a sampling rate of 100 Hz
+        is convolved with a 50 ms non-normalized gaussian kernel.
+
+            >>> tsd.smooth(50, time_units='ms', norm=False)
+
+        This line is equivalent to :
+
+            >>> from scipy.signal.windows import gaussian
+            >>> kernel = gaussian(M = 500, std=int(tsd.rate*0.05))
+            >>> tsd.convolve(window)
+
+        It is generally a good idea to visualize the kernel before applying any convolution.
+
+        See the scipy documentation for the [gaussian window](https://docs.scipy.org/doc/scipy/reference/generated/scipy.signal.windows.gaussian.html)
 
         Parameters
         ----------
-        std : int
-            Standard deviation
-        size : int
-            Description
+        std : Number
+            Standard deviation in units of time
+        time_units : str, optional
+            The time units in which std is specified ('us', 'ms', 's' [default]).
+        size_factor : int, optional
+            How long should be the kernel size as a function of the standard deviation. Default is 100.
+        norm : bool, optional
+            Whether to normalized the gaussian kernel or not. Default is `True`.
 
         Returns
         -------
         Tsd, TsdFrame, TsdTensor
             Time series convolved with a gaussian kernel
+
         """
-        assert isinstance(std, int), "std should be type int"
-        assert isinstance(size, int), "size should be type int"
-        window = signal.windows.gaussian(size, std=std)
-        window = window / window.sum()
+        assert isinstance(std, (int, float)), "std should be type int or float"
+        assert isinstance(size_factor, int), "size_factor should be of type int"
+        assert isinstance(norm, bool), "norm should be of type boolean"
+        assert isinstance(time_units, str), "time_units should be of type str"
+
+        std = TsIndex.format_timestamps(np.array([std]), time_units)[0]
+
+        std_size = int(self.rate * std)
+        M = std_size * size_factor
+        window = signal.windows.gaussian(M=M, std=std_size)
+
+        if norm:
+            window = window / window.sum()
+
         return self.convolve(window)
 
     def interpolate(self, ts, ep=None, left=None, right=None):
-        """Wrapper of the numpy linear interpolation method. See https://numpy.org/doc/stable/reference/generated/numpy.interp.html for an explanation of the parameters.
+        """Wrapper of the numpy linear interpolation method. See [numpy interpolate](https://numpy.org/doc/stable/reference/generated/numpy.interp.html)
+        for an explanation of the parameters.
         The argument ts should be Ts, Tsd, TsdFrame, TsdTensor to ensure interpolating from sorted timestamps in the right unit,
 
         Parameters